# docs/sms-api-doc.yaml
openapi: 3.0.1
info:
  title: School Management System API
  version: "1.3.3"
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /yearbooks:
    get:
      tags:
      - Yearbooks
      summary: Find all yearbooks
      description: "Retrieves all existing yearbooks in the system. Required authority\
        \ level: Administrator"
      operationId: getYearbooks
      responses:
        "401":
          description: Full authentication is required to access this resource
        "200":
          description: Returns existing yearbooks list (empty result as well)
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/YearbookSummaryResponse'
        "403":
          description: User lacks permission to access this resource
      security:
      - bearerAuth: []
    post:
      tags:
      - Yearbooks
      summary: Create yearbook
      description: "Adds new yearbook with unique symbol. Required authority level:\
        \ Administrator"
      operationId: addYearbook
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddYearbookRequest'
        required: true
      responses:
        "406":
          description: Subject not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/YearbookDetailedResponse'
        "201":
          description: Created successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/YearbookDetailedResponse'
        "404":
          description: Head teacher not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/YearbookDetailedResponse'
        "401":
          description: Full authentication is required to access this resource
        "403":
          description: User lacks permission to access this resource
        "409":
          description: Already exists for symbol or head teacher
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/YearbookDetailedResponse'
      security:
      - bearerAuth: []
...
  /yearbooks/{yearbookId}:
    get:
      tags:
      - Yearbooks
      summary: Find yearbook with ID
      description: "Retrieves existing yearbook with given ID. Required authority\
        \ level: Administrator, Teacher, Student"
      operationId: getYearbook
      parameters:
      - name: yearbookId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "404":
          description: Not found with ID
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/YearbookDetailedResponse'
        "401":
          description: Full authentication is required to access this resource
        "200":
          description: Returns existing yearbook
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/YearbookDetailedResponse'
        "403":
          description: User lacks permission to access this resource
      security:
      - bearerAuth: []
    delete:
      tags:
      - Yearbooks
      summary: Delete yearbook
      description: "Remove yearbook from the system if not associated with students.\
        \ Required authority level: Administrator"
      operationId: deleteYearbook
      parameters:
      - name: yearbookId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "204":
          description: Removed successfully
          content:
            '*/*':
              schema:
                type: string
        "406":
          description: Removal canceled due to existing yearbook association with
            students
          content:
            '*/*':
              schema:
                type: string
        "401":
          description: Full authentication is required to access this resource
        "404":
          description: Yearbook not found
          content:
            '*/*':
              schema:
                type: string
        "403":
          description: User lacks permission to access this resource
      security:
      - bearerAuth: []
    patch:
      tags:
      - Yearbooks
      summary: Update yearbook
      description: "Modifies yearbook data and subject association. Required authority\
        \ level: Administrator"
      operationId: updateYearbook
      parameters:
      - name: yearbookId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchYearbookRequest'
        required: true
      responses:
        "404":
          description: Head teacher not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/YearbookDetailedResponse'
        "200":
          description: Modification successful
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/YearbookDetailedResponse'
        "401":
          description: Full authentication is required to access this resource
        "403":
          description: User lacks permission to access this resource
      security:
      - bearerAuth: []
...
  /schedules/version/yearbook/{yearbookId}:
    get:
      tags:
      - Schedules
      summary: Get all schedule versions for yearbook
      description: "Retrieves all schedule versions for Yearbook that are not marked\
        \ as removed. Required authority level: Administrator"
      operationId: getScheduleVersionsForYearbook
      parameters:
      - name: yearbookId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Returns list of objects (empty result as well)
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScheduleVersionSummaryResponse'
        "401":
          description: Full authentication is required to access this resource
        "403":
          description: User lacks permission to access this resource
      security:
      - bearerAuth: []
...
  /occurrences/yearbook/{yearbookId}:
    get:
      tags:
      - Schedule Occurrences
      summary: Get schedule occurrences for yearbook
      description: "Retrieves all schedule (active) occurrences planned for given\
        \ Yearbook in particular time range. Time range is configured in app configuration\
        \ value. Required authority level: Administrator, Teacher, Student, Parent"
      operationId: getOccurrencesForYearbook
      parameters:
      - name: yearbookId
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "404":
          description: Required app config not found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScheduleOccurrenceSummaryResponse'
        "401":
          description: Full authentication is required to access this resource
        "406":
          description: Cannot fetch occurrences because there is no active schedule
            version
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScheduleOccurrenceSummaryResponse'
        "403":
          description: User lacks permission to access this resource
        "200":
          description: Returns objects list (empty result as well)
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScheduleOccurrenceSummaryResponse'
      security:
      - bearerAuth: []
...
components:
  schemas:
    AddYearbookRequest:
      required:
      - headTeacherId
      - mainCourseSubjectsIds
      - startingYear
      - symbol
      type: object
      properties:
        headTeacherId:
          type: integer
          format: int32
        symbol:
          type: string
        startingYear:
          type: string
          format: date-time
        graduationYear:
          type: string
          format: date-time
        mainCourseSubjectsIds:
          type: array
          items:
            type: integer
            format: int32
    StudentSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        birthDate:
          type: string
          format: date-time
        gender:
          type: string
          enum:
          - MALE
          - FEMALE
        phoneNumber:
          type: string
    SubjectSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        code:
          type: string
        description:
          type: string
        subjectType:
          type: string
    TeacherSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        birthDate:
          type: string
          format: date-time
        gender:
          type: string
          enum:
          - MALE
          - FEMALE
        phoneNumber:
          type: string
    YearbookDetailedResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        headTeacher:
          $ref: '#/components/schemas/TeacherSummaryResponse'
        symbol:
          type: string
        startingYear:
          type: string
          format: date-time
        graduationYear:
          type: string
          format: date-time
        studentsInYearbook:
          type: array
          items:
            $ref: '#/components/schemas/StudentSummaryResponse'
        mainCourseSubjects:
          type: array
          items:
            $ref: '#/components/schemas/SubjectSummaryResponse'
...
    YearbookSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        symbol:
          type: string
        headTeacher:
          $ref: '#/components/schemas/TeacherSummaryResponse'
        startingYear:
          type: string
          format: date-time
        graduationYear:
          type: string
          format: date-time
    LocalTime:
      type: object
      properties:
        hour:
          type: integer
          format: int32
        minute:
          type: integer
          format: int32
        second:
          type: integer
          format: int32
        nano:
          type: integer
          format: int32
...
    ScheduleVersionSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        active:
          type: boolean
...
    ScheduleOccurrenceSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        schedule:
          $ref: '#/components/schemas/ScheduleSummaryResponse'
        occurrenceDateTime:
          type: string
          format: date-time
        occurrenceEndTime:
          $ref: '#/components/schemas/LocalTime'
        status:
          type: string
          enum:
          - SCHEDULED
          - ONGOING
          - COMPLETED
          - CANCELLED
    ScheduleSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        dayOfWeek:
          type: string
          enum:
          - MONDAY
          - TUESDAY
          - WEDNESDAY
          - THURSDAY
          - FRIDAY
          - SATURDAY
          - SUNDAY
        startTime:
          $ref: '#/components/schemas/LocalTime'
        endTime:
          $ref: '#/components/schemas/LocalTime'
        classroomName:
          type: string
        teacherName:
          type: string
        subjectName:
          type: string
        scheduleVersionId:
          type: integer
          format: int32
        status:
          type: string
          enum:
          - SCHEDULED
          - CANCELLED
          - RESCHEDULED
          - DELETED
          - UPDATED
...
    PatchYearbookRequest:
      type: object
      properties:
        headTeacherId:
          type: integer
          format: int32
        symbol:
          type: string
        startingYear:
          type: string
          format: date-time
        targetGraduationYear:
          type: string
          format: date-time
        mainCourseSubjectsIdsToAdd:
          type: array
          items:
            type: integer
            format: int32
        mainCourseSubjectsIdsToRemove:
          type: array
          items:
            type: integer
            format: int32
...
  securitySchemes:
    bearerAuth:
      type: http
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT